pyproject.toml
[project]
name = "iwm-cot"
version = "0.1.0"
description = "Interpreter-Grounded Working-Memory CoT experiments"
requires-python = ">=3.10"

[tool.uv]

[tool.poetry.dependencies]
python = ">=3.10"
torch = "^2.2.0"
transformers = "^4.44.0"
datasets = "^2.20.0"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
wandb = "^0.16.0"
optuna = "^3.6.0"
matplotlib = "^3.9.0"
seaborn = "^0.13.2"
pyyaml = "^6.0.2"
scipy = "^1.11.0"
pandas = "^2.2.0"
